+++ examples/stats.rs
@@ -9,7 +9,10 @@ fn main() {
     }
 
     let text = std::fs::read_to_string(&args[1]).unwrap();
-    let opt = roxmltree::ParsingOptions { allow_dtd: true };
+    let opt = roxmltree::ParsingOptions {
+        allow_dtd: true,
+        nodes_limit: 0,
+    };
     let doc = match roxmltree::Document::parse_with_options(&text, opt) {
         Ok(v) => v,
         Err(e) => {